third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/claim-worker-fetch.https-expected.txt

This is a testharness.js-based test.
PASS fetch() in Worker should be intercepted after the client is claimed.
FAIL fetch() in nested Worker should be intercepted after the client is claimed. assert_equals: fetch() should not be intercepted. expected "a simple text file\n" but got "Fail: undefined"
Harness: the test ran to completion.
